\name{aode}
\alias{aode}
\title{
AODE
}
\description{
Implementacja fazy trenującej algorytmu AODE. Traktuje wartości numeryczne jako 
dyskretne.
}
\usage{
aode(x, ...)
}
\arguments{
  \item{x}{
Macierz atrybutów. Nie zawiera atrybuty klasy.
}
  \item{y}{
Wektor klas dla każdego z przykładów z \code{x}.  
}
  \item{laplace}{
Współczynnik wygładzenia Laplace'a.
}
  \item{\dots}{
Pozostałe argumenty
}
}
\details{
}
\value{
    \item{vfreq } {Description...}
    \item{tables1 } {Description...}
    \item{tables2 } {Description...}
    \item{levels } {Description...}
    \item{call } {Description...}
    \item{laplace } {Description...}
}
\references{
}
\author{
}
\note{
}
\seealso{
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.
}
\keyword{methods}
