\name{bayes-package}
\alias{bayes-package}
\alias{bayes}
\docType{package}
\title{
Bayes - pakiet stworzony w ramach zaliczenia projektu z przedmiotu MOW.
}
\description{
Implementacja standardowego klasyfikatora Bayesowskiego oraz dwóch jego
modyfikacji : AODE(ang. Averaged One-Dependence Estimators), TAN(ang. Tree 
Augmented Naive Bayes).
}
\details{
\tabular{ll}{
Package: \tab bayes\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-05-28\cr
License: \tab MIT\cr
}
}
\author{
Michał Dębski, Cezary Pawłowski, Przemysław Piórkowski
}
\references{

}
\keyword{ package }
\examples{
    install.packages("/path/to/package/bayes_1.0.tar.gz", repos = NULL, type="source")
}
