% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesAnalysis_yepFormat.R
\name{BayesAnalysis}
\alias{BayesAnalysis}
\title{A CLDL Function}
\usage{
BayesAnalysis(data, hypotheses, list_id_vars, proportion = FALSE,
  number_of_subjects = NULL, levels = NULL)
}
\arguments{
\item{hypotheses}{the list of hypotheses (as strings)}

\item{list_id_vars}{the list of independent variables}

\item{number_of_subjects}{default is NULL, value required if ggplotting by proportion}

\item{levels}{default is NULL, if !NULL it is a dataframe/csv with levels (hypothesis-predictions). function will combine levels and data files and create 'yeps' and then add randomYep column to data. else, assumes this is already complete and yep columns already exist and analysis applied from that point}

\item{df}{the dataframe, must contain [hypothesis]Yep columns to indicate if response was consistent with each hypothesis. Be sure to include "randomYep" colunn if you want chance (default 50\%) as a hypothesis alternative.}

\item{by_proportion}{defaults to FALSE: ggplot by number of subjects, if TRUE, plots by proportion of subjects}

\item{include_random}{default is TRUE, includes "random" as hypothesis alternative, and chance defualts to 50\% [change this in calculateLikelihood fxn]}
}
\value{
returns 'kdata' containing the k values, 'data3' containing summary and winner, 'data4' which collapses by block, 'data4.long' which melts to long format for graphing, and the Bayes plot
}
\description{
This function applies our Bayes Analysis to data (in YepFormat)
}
\examples{
BayesAnalysis_yepFormat(data, c("strong", "weak", "average", "X1biggest", "X2biggest"), c("subject", "mixing", "alignment"))
BayesAnalysis_yepFormat(data, c("strong", "weak", "average", "X1biggest", "X2biggest"), c("subject", "mixing", "alignment"), TRUE, 30)
BayesAnalysis_yepFormat(rawdata, c("strong", "weak", "average"), c("subject", "mixing", "alignment"), levels=read.csv("~/Desktop/lst2-levels-formerging.csv"))

}
\keyword{Bayes}
\keyword{Kvalue}
\keyword{Likelihood}

