% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesAnalysis_HelperFunctions.R
\name{BayesPlot}
\alias{BayesPlot}
\title{A CLDL Function}
\usage{
BayesPlot(df.long, list_id_vars, by_proportion = FALSE,
  number_of_subjects = NULL)
}
\arguments{
\item{df.long}{the long dataframe for plotting, with k values for each hypothesis and summary by hypothesis}

\item{list_id_vars}{the independent variables}

\item{by_proportion}{defaults to FALSE and plots by number, if TRUE, plots by proportion}

\item{number_of_subjects}{defaults to NULL, required to calculate proportions if plotting by proportion}
}
\value{
returns ggplot
}
\description{
This function generates our Bayes Plot to conclude our analysis
}
\examples{
BayesPlot(data4.long, c("subject", "mixing", "alignment"), TRUE, 30)
BayesPlot(data4.long, c("subject", "mixing", "alignment"), FALSE)
BayesPlot(data4.long, c("subject", "mixing", "alignment"))

}
\keyword{Bayes}

