% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesAnalysis_HelperFunctions.R
\name{calculateLikelihood}
\alias{calculateLikelihood}
\title{A CLDL Function}
\usage{
calculateLikelihood(data, list_id_vars, measure_var_yep, measure_var_nope,
  hypothesis)
}
\arguments{
\item{data}{the dataset; each hypothesis requires a column with subject response consistent with hypothesis = 1, else 0; and a column for response inconsistent=1, else 0.}

\item{list_id_vars}{the list of independent variables for the recast function (list of strings)}

\item{measure_var_yep}{colname of 1s/0s corresponding to hypothesis-CONSISTENT (yep) responses (as string)}

\item{measure_var_nope}{colname of 1s/0s corresponding to hypothesis-INCONSISTENT (nope) responses (as string)}
}
\value{
returns complete dataset to global environment
}
\description{
This function calculates the likelihoods of a hypothesis for use in our Bayes Analysis
}
\examples{
calculateLikelihood(data, c("subject", "mixing", "alignment"), "strongYep", "strongNope", "Strong")
calculateLikelihood(data, c("subject", "mixing", "alignment"), "weakYep", "weakNope", "Weak")

}
\keyword{Bayes}
\keyword{likelihood}

