% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesAnalysis_HelperFunctions.R
\name{makeYeps_fromLevels}
\alias{makeYeps_fromLevels}
\title{A CLDL Function}
\usage{
makeYeps_fromLevels(df, levels, hypotheses)
}
\arguments{
\item{df}{the dataframe}

\item{levels}{the csv containing levels (must contain "[hypothesis]px" column for every hypothesis)}

\item{hypotheses}{a list of the hypotheses (must each have a "px" column in the levels file)}
}
\value{
returns 'data' dataframe to Global Env., containing the Yep value columns (value "correct" by hypothesis)
}
\description{
This function generates data in "Yep format": compares the hypothesis-predictions in the levels file to subject responses and writes "yep" columns to data to indicate if response was hypothesis-consistent
}
\examples{
makeYeps_fromLevels(data, levels, hyps)

}

