\name{print.aode}
\alias{print.aode}
AODE
\title{
Wypisanie elementów modelu algorytmu AODE.
}
\description{
}
\usage{
print.aode(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
Model
}
  \item{\dots}{
Pozostałe argumenty
}
}
\details{
}
\value{
}
\references{
}
\author{
}
\note{
}
\seealso{
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.
}
\keyword{methods}
