% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/R commands.R
\name{Rcommands}
\alias{Rcommands}
\title{R commands}
\usage{
Rcommands()
}
\arguments{
\item{nothing}{There are no parameters for this function}
}
\value{
This function doesn't return any data
}
\description{
This script analyses the sutdent data and shows how to uses plots, histograms, and fit an abline fit on the data
}
\note{
This is just a function for demonstration
}

